OpenMP performance and profiling tools (#407)
* Update docs and re-add covid runner

* Docs

* More notes

* Experimenting

* Docs on profiling

* Added flame graph task

* Fix symbol replacement issue

* Flame graph sub prefix

* More attempts to debug covid stuff

* Fixed bug in flame graph gen

* Profiling

* Formatting

* Vector instructions

* Profiling and main-threading

* Extra tracing

* Correct timer name

* Removed some unused branches of our openmp implementation

* Tidying up

* Tidy-up

* Latest cpp

* Fixing up tests

* Revert to master faabric version

* Mayfail on MPI async test

* Actually comment out async MPI test

* Update dynamic module offsets

+++ src/wavm/openmp.cpp
@@ -22,14 +22,15 @@ namespace wasm {
 // ------------------------------------------------
 
 #define OMP_FUNC(str)                                                          \
-    auto ctx = threads::getOpenMPContext();                                    \
+    threads::OpenMPContext& ctx = threads::getOpenMPContext();                 \
     auto logger = faabric::util::getLogger();                                  \
-    logger->trace("OMP {}: " str, ctx.threadNumber);
+    logger->trace("OMP {} ({}): " str, ctx.threadNumber, ::gettid());
 
 #define OMP_FUNC_ARGS(formatStr, ...)                                          \
-    auto ctx = threads::getOpenMPContext();                                    \
+    threads::OpenMPContext& ctx = threads::getOpenMPContext();                 \
     auto logger = faabric::util::getLogger();                                  \
-    logger->trace("OMP {}: " formatStr, ctx.threadNumber, __VA_ARGS__);
+    logger->trace(                                                             \
+      "OMP {} ({}): " formatStr, ctx.threadNumber, gettid(), __VA_ARGS__);
 
 // ------------------------------------------------
 // THREAD NUMS AND LEVELS
@@ -352,9 +353,11 @@ WAVM_DEFINE_INTRINSIC_FUNCTION(env,
 // ----------------------------------------------------
 
 /**
- * The "real" version of this function is implemented in the openmp source at
- * openmp/runtime/src/kmp_csupport.cpp. This in turn calls __kmp_fork_call which
- * does the real heavy lifting (see openmp/runtime/src/kmp_runtime.cpp)
+ * The "real" version of this function is implemented in the openmp source at:
+ * https://github.com/llvm/llvm-project/blob/main/openmp/runtime/src/kmp_csupport.cpp
+ *
+ * It calls into __kmp_fork call to do most of the work, which is here:
+ * https://github.com/llvm/llvm-project/blob/main/openmp/runtime/src/kmp_runtime.cpp
  *
  * @param locPtr pointer to the source location info (type ident_t)
  * @param argc number of arguments to pass to the microtask
@@ -379,7 +382,6 @@ WAVM_DEFINE_INTRINSIC_FUNCTION(env,
     OMP_FUNC_ARGS(
       "__kmpc_fork_call {} {} {} {}", locPtr, argc, microtaskPtr, argsPtr);
 
-    auto conf = faabric::util::getSystemConfig();
     auto& sch = faabric::scheduler::getScheduler();
 
     WAVMWasmModule* parentModule = getExecutingWAVMModule();
@@ -390,32 +392,49 @@ WAVM_DEFINE_INTRINSIC_FUNCTION(env,
     std::shared_ptr<threads::Level> parentLevel = ctx.level;
     int nextNumThreads = parentLevel->getMaxThreadsAtNextLevel();
 
+    // Check if we're only doing single-threaded
+    bool isSingleThread = nextNumThreads == 1;
+
     // Set up the next level
     auto nextLevel = std::make_shared<threads::Level>(nextNumThreads);
     nextLevel->fromParentLevel(parentLevel);
 
     // Take memory snapshot
-    std::string snapshotKey = parentModule->snapshot(false);
-    logger->debug("Created OpenMP snapshot: {}", snapshotKey);
+    std::string snapshotKey;
+    if (!isSingleThread) {
+        snapshotKey = parentModule->snapshot(false);
+        logger->debug("Created OpenMP snapshot: {}", snapshotKey);
+    } else {
+        logger->debug("Not creating OpenMP snapshot for single thread");
+    }
 
     const faabric::Message* originalCall = getExecutingCall();
     const std::string origStr =
       faabric::util::funcToString(*originalCall, false);
 
-    // Get pointers to shared variables in host memory
+    // Prepare arguments for main thread and all others
+    std::vector<IR::UntaggedValue> mainArguments = { 0, argc };
     U32* sharedVarsPtr = nullptr;
     if (argc > 0) {
         sharedVarsPtr = Runtime::memoryArrayPtr<U32>(memoryPtr, argsPtr, argc);
     }
 
     // Set up the chained calls
-    std::vector<std::shared_ptr<faabric::Message>> msgs;
-    std::vector<int> callIds;
+    // Note that the main OpenMP thread is always executed in this thread, hence
+    // it is not scheduled remotely
+    std::shared_ptr<faabric::Message> masterMsg = nullptr;
+    std::vector<std::shared_ptr<faabric::Message>> childThreadMsgs;
+    std::vector<int> remoteCallIds;
     for (int threadNum = 0; threadNum < nextNumThreads; threadNum++) {
         // Create basic call
         std::shared_ptr<faabric::Message> call =
           faabric::util::messageFactoryShared(originalCall->user(),
                                               originalCall->function());
+
+        // Record master message if necessary
+        if (threadNum == 0) {
+            masterMsg = call;
+        }
 
         // All calls are async by definition
         call->set_isasync(true);
@@ -429,6 +448,7 @@ WAVM_DEFINE_INTRINSIC_FUNCTION(env,
         // Args
         for (int i = 0; i < argc; i++) {
             call->add_ompfunctionargs(sharedVarsPtr[i]);
+            mainArguments.emplace_back(sharedVarsPtr[i]);
         }
 
         call->set_ompthreadnum(threadNum);
@@ -438,40 +458,73 @@ WAVM_DEFINE_INTRINSIC_FUNCTION(env,
         call->set_ompeffdepth(nextLevel->activeLevels);
         call->set_ompmal(nextLevel->maxActiveLevels);
 
-        msgs.push_back(call);
+        if (threadNum > 0) {
+            childThreadMsgs.push_back(call);
+        }
     }
 
-    // Set up the request
-    faabric::BatchExecuteRequest req = faabric::util::batchExecFactory(msgs);
-    req.set_type(faabric::BatchExecuteRequest::THREADS);
-
-    // Submit it
-    std::vector<std::string> executedHosts = sch.callFunctions(req);
-
-    // Iterate through messages and see which need to be executed locally
+    // Attempt to schedule across hosts
     std::vector<std::future<int32_t>> localFutures;
-    for (int i = 0; i < executedHosts.size(); i++) {
-        std::string host = executedHosts.at(i);
-        bool isLocal = host.empty();
+    if (!isSingleThread) {
+        // Set up the request
+        faabric::BatchExecuteRequest req =
+          faabric::util::batchExecFactory(childThreadMsgs);
 
-        std::shared_ptr<faabric::Message> msg = msgs.at(i);
-        uint32_t msgId = msg->id();
+        req.set_type(faabric::BatchExecuteRequest::THREADS);
 
-        if (!isLocal) {
-            // Function is being executed remotely
-            logger->debug("Waiting for remote thread for call ID {}", msgId);
-            callIds.push_back(msgId);
-            continue;
+        // Submit it
+        std::vector<std::string> executedHosts = sch.callFunctions(req);
+
+        // Iterate through messages and see which need to be executed in local
+        // threads
+        for (int i = 0; i < executedHosts.size(); i++) {
+            std::string host = executedHosts.at(i);
+            bool isLocal = host.empty();
+
+            std::shared_ptr<faabric::Message> msg = childThreadMsgs.at(i);
+            uint32_t msgId = msg->id();
+
+            if (isLocal) {
+                logger->debug("Dispatching local task for OpenMP call ID {}",
+                              msgId);
+
+                // We execute the main thread in this thread so index is one
+                // higher
+                int threadIdx = i + 1;
+
+                // Execute the local thread
+                threads::OpenMPTask t(parentCall, msg, nextLevel, threadIdx);
+                localFutures.emplace_back(parentModule->executeOpenMPTask(t));
+            } else {
+                // Function is being executed remotely
+                logger->debug("Waiting for remote thread for call ID {}",
+                              msgId);
+                remoteCallIds.push_back(msgId);
+                continue;
+            }
         }
+    }
 
-        logger->debug("Executing local OMP thread for call ID {}", msgId);
-        threads::OpenMPTask t(parentCall, msg, nextLevel, i);
-        localFutures.emplace_back(parentModule->executeOpenMPTask(t));
+    // Execute the master task (just invoke the microtask directly)
+    IR::UntaggedValue mainThreadResult;
+    {
+        // We have to set up the context for the thread
+        setUpOpenMPContext(0, nextLevel);
+
+        // Execute the task
+        logger->debug("Executing master OMP thread");
+        WAVM::Runtime::Function* microtaskFunc =
+          parentModule->getFunctionFromPtr(microtaskPtr);
+        parentModule->executeFunction(
+          microtaskFunc, mainArguments, mainThreadResult);
+
+        // Now we reset the context for this main thread
+        setUpOpenMPContext(0, parentLevel);
     }
 
     // Await the results of all the remote threads
     int numErrors = 0;
-    for (auto callId : callIds) {
+    for (auto callId : remoteCallIds) {
         int callTimeoutMs = faabric::util::getSystemConfig().chainedCallTimeout;
 
         logger->info(
@@ -505,20 +558,26 @@ WAVM_DEFINE_INTRINSIC_FUNCTION(env,
         }
     }
 
+    if (mainThreadResult.i32 > 0) {
+        throw std::runtime_error("Master OpenMP thread failed");
+    }
+
     // Reset parent level for next setting of threads
     parentLevel->pushedThreads = -1;
 
     // Delete the snapshot from registered hosts
-    PROF_START(BroadcastDeleteSnapshot)
-    sch.broadcastSnapshotDelete(*originalCall, snapshotKey);
-    PROF_END(BroadcastDeleteSnapshot)
+    if (!isSingleThread) {
+        PROF_START(BroadcastDeleteSnapshot)
+        sch.broadcastSnapshotDelete(*originalCall, snapshotKey);
+        PROF_END(BroadcastDeleteSnapshot)
 
-    // Delete the snapshot locally
-    PROF_START(DeleteSnapshot)
-    faabric::snapshot::SnapshotRegistry& reg =
-      faabric::snapshot::getSnapshotRegistry();
-    reg.deleteSnapshot(snapshotKey);
-    PROF_END(DeleteSnapshot)
+        // Delete the snapshot locally
+        PROF_START(DeleteSnapshot)
+        faabric::snapshot::SnapshotRegistry& reg =
+          faabric::snapshot::getSnapshotRegistry();
+        reg.deleteSnapshot(snapshotKey);
+        PROF_END(DeleteSnapshot)
+    }
 }
 
 // -------------------------------------------------------
@@ -554,7 +613,7 @@ void for_static_init(I32 schedule,
     typedef typename std::make_unsigned<T>::type UT;
 
     auto logger = faabric::util::getLogger();
-    auto ctx = threads::getOpenMPContext();
+    threads::OpenMPContext& ctx = threads::getOpenMPContext();
 
     if (ctx.level->numThreads == 1) {
         *lastIter = true;
@@ -758,50 +817,9 @@ WAVM_DEFINE_INTRINSIC_FUNCTION(env,
 // REDUCTION
 // ---------------------------------------------------
 
-// Allowed reduce return values.
-// The single thread value is only used when there's one thread.
-// Multi threads signifies that multiple threads are operating
-// We don't currently support atomic reduce, but it signifies that the final
-// reduction must be performed atomically.
-enum ReduceReturnValue
-{
-    SINGLE_THREAD = 0,
-    MULTIPLE_THREADS = 1,
-    ATMOIC_REDUCE = 2,
-};
-
-/*
- * This method should return one of the reduce return values as detailed above.
- * AFAICT most of the arguments are not needed in our case.
- *
- * The reduction is performed by the master thread, so we must ensure that the
- * other threads have finished before this takes place.
- *
- * To interrogate where intermediate results are stored, we can look at the
- * reduce data and reduce func.
- *
- * The reduceFunc provides the reduction operation on two operands and returns
- * the result in its first argument. The function is of type:
- * void(*)(void *lhsData, void *rhsData)
- *
- * For more info, see OpenMP source at
- * https://github.com/llvm/llvm-project/blob/main/openmp/runtime/src/kmp_csupport.cpp
- */
 int reduceFinished()
 {
-    auto logger = faabric::util::getLogger();
-    auto ctx = threads::getOpenMPContext();
-
-    if (ctx.level->numThreads == 1) {
-        return SINGLE_THREAD;
-    }
-
-    // Notify the master thread that we've done our reduction
-    if (ctx.threadNumber != 0) {
-        ctx.level->masterWait(ctx.threadNumber);
-    }
-
-    return MULTIPLE_THREADS;
+    return 1;
 }
 
 /**
@@ -811,24 +829,23 @@ int reduceFinished()
  */
 void finaliseReduce(bool barrier)
 {
-    auto logger = faabric::util::getLogger();
-    auto ctx = threads::getOpenMPContext();
-
-    // Master must make sure all other threads are done
-    if (ctx.threadNumber == 0) {
-        ctx.level->masterWait(ctx.threadNumber);
-
-        logger->debug("Master thread finished reduce");
-    }
+    threads::OpenMPContext& ctx = threads::getOpenMPContext();
 
     // Everyone waits if there's a barrier
     if (barrier) {
+        PROF_START(FinaliseReduceBarrier)
         ctx.level->barrier.wait();
+        PROF_END(FinaliseReduceBarrier)
     }
 }
 
 /**
- * A blocking reduce that includes an implicit barrier.
+ * It seems that in our case, always returning 1 for both kmpc_reduce and
+ * kmpc_reduce_nowait gets the right result.
+ *
+ * In the OpenMP source we can see a more varied set of return values, but these
+ * are for cases we don't yet support:
+ * https://github.com/llvm/llvm-project/blob/main/openmp/runtime/src/kmp_csupport.cpp
  */
 WAVM_DEFINE_INTRINSIC_FUNCTION(env,
                                "__kmpc_reduce",
@@ -851,12 +868,11 @@ WAVM_DEFINE_INTRINSIC_FUNCTION(env,
                   reduceFunc,
                   lockPtr);
 
-    return reduceFinished();
+    return 1;
 }
 
 /**
- * The nowait version is used for a reduce clause with the nowait argument, or
- * when the parallel section finished straight after the reduce.
+ * See __kmpc_reduce
  */
 WAVM_DEFINE_INTRINSIC_FUNCTION(env,
                                "__kmpc_reduce_nowait",
@@ -879,7 +895,7 @@ WAVM_DEFINE_INTRINSIC_FUNCTION(env,
                   reduceFunc,
                   lockPtr);
 
-    return reduceFinished();
+    return 1;
 }
 
 /**
@@ -894,7 +910,6 @@ WAVM_DEFINE_INTRINSIC_FUNCTION(env,
                                I32 lck)
 {
     OMP_FUNC_ARGS("__kmpc_end_reduce {} {} {}", loc, gtid, lck);
-
     finaliseReduce(true);
 }
 
@@ -910,7 +925,6 @@ WAVM_DEFINE_INTRINSIC_FUNCTION(env,
                                I32 lck)
 {
     OMP_FUNC_ARGS("__kmpc_end_reduce_nowait {} {} {}", loc, gtid, lck);
-
     finaliseReduce(false);
 }
